Replace glbinding with Glad. Improve project structure.
The idea behind glbinding makes a lot of sense: type-safety for the
OpenGL C/C++ API. Sadly, the execution is lacking - glbinding is complex,
takes a long time to compile and causes trouble with the build system.
Glad is leightweight, easy to integrate and provides OpenGL access in a
way most people are familiar with. For finding type errors, we instead
rely on Glad's glGetError() post function hook and OpenGL debug output.
It's a tradeoff, but I think it's worth it.