Fix issue with NaN occurance.
The problem is that grid cells are limited to 256 particles. If the
count is higher than that, particles may be lost. When that happens, it
cannot find itself in the density neighborhood search. And when that
happens, we divide by 0.0 in the sim step 2, leading to NaNs.

The correct way to fix this would be adjusting the SPH parameters so
that this never happens. However if someone plays around with these or
they get changed without care, this issue might occur. This 'quickfix'
for hiding the symptoms is very cheap.