Fix velocity and position integration.
Before this commit, we incorrectlty integrated the particle velocity and
position while still using them in the second simulation step. Now, we
sample a velocity 3D texture with HW filtering and update the velocity
in the same shader. The position integration has been moved to the next
step, right before the grid construction. I've also adjusted some SPH
paramaters and increased the default particle count to 50000.